Add permissions checking to server

+++ src/server/handler.go
@@ -17,7 +17,7 @@ import (
 type FileSystem struct {
 	directory   string
 	uuid        string
-	permissions map[string]string
+	permissions []string
 	lock        sync.Mutex
 }
 
@@ -27,8 +27,9 @@ type FileSystem struct {
 func CreateHandler(base string, perm *ssh.Permissions) sftp.Handlers {
 	p := FileSystem{
 		//directory: path.Join(base, perm.Extensions["uuid"]),
-		directory: "/Users/dane/Downloads",
-		uuid:      perm.Extensions["uuid"],
+		directory:   "/Users/dane/Downloads",
+		uuid:        perm.Extensions["uuid"],
+		permissions: strings.Split(perm.Extensions["permissions"], ","),
 	}
 
 	return sftp.Handlers{
@@ -41,6 +42,13 @@ func CreateHandler(base string, perm *ssh.Permissions) sftp.Handlers {
 
 // Creates a reader for a file on the system and returns the reader back.
 func (fs FileSystem) Fileread(request *sftp.Request) (io.ReaderAt, error) {
+	// Check first if the user can actually open and view a file. This permission is named
+	// really poorly, but it is checking if they can read. There is an addition permission,
+	// "save-files" which determines if they can write that file.
+	if !fs.can("edit-files") {
+		return nil, sftp.ErrSshFxPermissionDenied
+	}
+
 	path, err := fs.buildPath(request.Filepath)
 	if err != nil {
 		return nil, sftp.ErrSshFxNoSuchFile
@@ -74,6 +82,13 @@ func (fs FileSystem) Filewrite(request *sftp.Request) (io.WriterAt, error) {
 	// If the file doesn't exist we need to create it, as well as the directory pathway
 	// leading up to where that file will be created.
 	if os.IsNotExist(statErr) {
+		// This is a different pathway than just editing an existing file. If it doesn't exist already
+		// we need to determine if this user has permission to create files.
+		if !fs.can("create-files") {
+			return nil, sftp.ErrSshFxPermissionDenied
+		}
+
+		// Create all of the directories leading up to the location where this file is being created.
 		if err := os.MkdirAll(filepath.Dir(path), 0755); err != nil {
 			logger.Get().Errorw("error making path for file",
 				zap.String("source", path),
@@ -98,6 +113,12 @@ func (fs FileSystem) Filewrite(request *sftp.Request) (io.WriterAt, error) {
 	// If we've made it here it means the file already exists and we don't need to do anything
 	// fancy to handle it. Just pass over the request flags so the system knows what the end
 	// goal with the file is going to be.
+	//
+	// But first, check that the user has permission to save modified files.
+	if !fs.can("save-files") {
+		return nil, sftp.ErrSshFxPermissionDenied
+	}
+
 	file, err := os.OpenFile(path, int(request.Flags), 0644)
 	if err != nil {
 		logger.Get().Errorw("error writing to existing file",
@@ -135,6 +156,10 @@ func (fs FileSystem) Filecmd(request *sftp.Request) error {
 	case "SetStat", "Setstat":
 		return sftp.ErrSshFxOpUnsupported
 	case "Rename":
+		if !fs.can("move-files") {
+			return sftp.ErrSshFxPermissionDenied
+		}
+
 		if err := os.Rename(path, target); err != nil {
 			logger.Get().Errorw("failed to rename file",
 				zap.String("source", path),
@@ -146,6 +171,10 @@ func (fs FileSystem) Filecmd(request *sftp.Request) error {
 
 		return sftp.ErrSshFxOk
 	case "Rmdir":
+		if !fs.can("delete-files") {
+			return sftp.ErrSshFxPermissionDenied
+		}
+
 		if err := os.RemoveAll(path); err != nil {
 			logger.Get().Errorw("failed to remove directory", zap.String("source", path), zap.Error(err))
 			return sftp.ErrSshFxFailure
@@ -153,6 +182,10 @@ func (fs FileSystem) Filecmd(request *sftp.Request) error {
 
 		return sftp.ErrSshFxOk
 	case "Mkdir":
+		if !fs.can("create-files") {
+			return sftp.ErrSshFxPermissionDenied
+		}
+
 		if err := os.MkdirAll(path, 0755); err != nil {
 			logger.Get().Errorw("failed to create directory", zap.String("source", path), zap.Error(err))
 			return sftp.ErrSshFxFailure
@@ -160,6 +193,10 @@ func (fs FileSystem) Filecmd(request *sftp.Request) error {
 
 		return sftp.ErrSshFxOk
 	case "Symlink":
+		if !fs.can("create-files") {
+			return sftp.ErrSshFxPermissionDenied
+		}
+
 		if err := os.Symlink(path, target); err != nil {
 			logger.Get().Errorw("failed to create symlink",
 				zap.String("source", path),
@@ -171,6 +208,10 @@ func (fs FileSystem) Filecmd(request *sftp.Request) error {
 
 		return sftp.ErrSshFxOk
 	case "Remove":
+		if !fs.can("delete-files") {
+			return sftp.ErrSshFxPermissionDenied
+		}
+
 		if err := os.Remove(path); err != nil {
 			logger.Get().Errorw("failed to remove a file", zap.String("source", path), zap.Error(err))
 			return sftp.ErrSshFxFailure
@@ -192,6 +233,10 @@ func (fs FileSystem) Filelist(request *sftp.Request) (sftp.ListerAt, error) {
 
 	switch request.Method {
 	case "List":
+		if !fs.can("list-files") {
+			return nil, sftp.ErrSshFxPermissionDenied
+		}
+
 		files, err := ioutil.ReadDir(path)
 		if err != nil {
 			logger.Get().Error("error listing directory", zap.Error(err))
@@ -200,6 +245,10 @@ func (fs FileSystem) Filelist(request *sftp.Request) (sftp.ListerAt, error) {
 
 		return ListerAt(files), nil
 	case "Stat":
+		if !fs.can("list-files") {
+			return nil, sftp.ErrSshFxPermissionDenied
+		}
+
 		file, err := os.Open(path)
 		defer file.Close()
 
@@ -215,14 +264,12 @@ func (fs FileSystem) Filelist(request *sftp.Request) (sftp.ListerAt, error) {
 		}
 
 		return ListerAt([]os.FileInfo{s}), nil
+	default:
 		// Before adding readlink support we need to evaluate any potential security risks
 		// as a result of navigating around to a location that is outside the home directory
 		// for the logged in user. I don't forsee it being much of a problem, but I do want to
 		// check it out before slapping some code here. Until then, we'll just return an
 		// unsupported response code.
-		//
-		// case "Readlink":
-	default:
 		return nil, sftp.ErrSshFxOpUnsupported
 	}
 }
@@ -231,11 +278,36 @@ func (fs FileSystem) Filelist(request *sftp.Request) (sftp.ListerAt, error) {
 // from their data directory. After normalization if the directory is still within their home
 // path it is returned. If they managed to "escape" an error will be returned.
 func (fs FileSystem) buildPath(rawPath string) (string, error) {
+	// Calling filepath.Clean on the joined directory will resolve it to the absolute path,
+	// removing any ../ type of path resolution, and leaving us with the absolute final path.
 	path := filepath.Clean(filepath.Join(fs.directory, rawPath))
 
+	// If the new path doesn't start with their root directory there is clearly an escape
+	// attempt going on, and we should NOT resolve this path for them.
 	if !strings.HasPrefix(path, fs.directory) {
 		return "", errors.New("invalid path resolution")
 	}
 
 	return path, nil
+}
+
+// Determines if a user has permission to perform a specific action on the SFTP server. These
+// permissions are defined and returned by the Panel API.
+func (fs FileSystem) can(permission string) bool {
+	// Server owners and super admins have their permissions returned as '[*]' via the Panel
+	// API, so for the sake of speed do an initial check for that before iterating over the
+	// entire array of permissions.
+	if len(fs.permissions) == 1 && fs.permissions[0] == "*" {
+		return true
+	}
+
+	// Not the owner or an admin, loop over the permissions that were returned to determine
+	// if they have the passed permission.
+	for _, p := range fs.permissions {
+		if p == permission {
+			return true
+		}
+	}
+
+	return false
 }
